Moves SQL generation tests to unit tests
