+++ GitVersion.yml
@@ -6,7 +6,6 @@ mode: ContinuousDelivery
 legacy-semver-padding: 0
 build-metadata-padding: 0
 commits-since-version-source-padding: 0
-next-version: 1.0
 
 branches: 
   master:
