--- tests/GenericSearch.IntegrationTests/Properties/AssemblyGlobals.cs
@@ -1,3 +1,0 @@
-﻿using System.Diagnostics.CodeAnalysis;
-
-[assembly:ExcludeFromCodeCoverage]
