+++ tests/GenericSearch.UnitTests/Data/Entities/Author.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace GenericSearch.IntegrationTests.Internal.Data.Entities
+namespace GenericSearch.UnitTests.Data.Entities
 {
     public class Author
     {
