+++ tests/GenericSearch.UnitTests/Data/Entities/Post.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace GenericSearch.IntegrationTests.Internal.Data.Entities
+namespace GenericSearch.UnitTests.Data.Entities
 {
     public class Post
     {
